separated into own components per functionality
